refactor: replace double with BigDecimal

+++ src/main/java/com/paymentteamproject/domain/payment/dto/StartPaymentRequest.java
@@ -3,11 +3,13 @@ package com.paymentteamproject.domain.payment.dto;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class StartPaymentRequest {
     @NotNull(message = "id는 필수 값 입니다.")
     private Long orderId;
     @NotNull(message = "총금액은 필수 값 입니다.")
-    private double totalAmount;
-    private double pointsToUse;
+    private BigDecimal totalAmount;
+    private BigDecimal pointsToUse;
 }
